Implemented screen scrolling in all directions, removed the invisible floor, now if you fall off, you will just keep falling, as the "world" is 'infinite' now.
Known bugs that should exist: if you have a wall on your right, and you are on the left side of the screen, you will probably currently be able to use the screen scroll to phase through the block.

The 'object' class in the html is what you need to add to anything that you want to be screenshifted. (so everything in the game)

the 'brick' class in the html is what you need to add to anything that you want to have collisions. (As of now)